refactored layers, deleted local db
