--- lib/features/movies/application/services/interface/i_service.dart
@@ -1,3 +1,0 @@
-abstract class IService<T> {
-  Future<T> call({params});
-}
