--- lib/features/movies/data/datasources/local/converters/genres_list_converter.dart
@@ -1,18 +1,0 @@
-import 'dart:convert';
-
-import 'package:floor/floor.dart';
-
-class GenresListConverter extends TypeConverter<List<int>, String> {
-  @override
-  List<int> decode(String databaseValue) {
-    final dynamic jsonFile = json.decode(databaseValue);
-    return List<int>.from(jsonFile['Genres']);
-  }
-
-  @override
-  String encode(List<int> value) {
-    final Map<String, dynamic> data = <String, dynamic>{};
-    data.addAll(<String, dynamic>{'Genres': value});
-    return json.encode(data);
-  }
-}
