--- lib/features/movies/data/datasources/local/converters/vote_average_converter.dart
@@ -1,13 +1,0 @@
-import 'package:floor/floor.dart';
-
-class VoteAverageConverter extends TypeConverter<num, String> {
-  @override
-  num decode(String databaseValue) {
-    return num.parse(databaseValue);
-  }
-
-  @override
-  String encode(num value) {
-    return value.toString();
-  }
-}
