--- lib/features/movies/data/datasources/local/database_initializer.dart
@@ -1,13 +1,0 @@
-import '../../../../../constants/strings.dart';
-import 'movies_database.dart';
-
-class DatabaseInitializer {
-  Future<MoviesDatabase> initializeDB() async {
-    MoviesDatabase databaseInstance;
-    databaseInstance = await $FloorMoviesDatabase
-        .databaseBuilder(Strings.databaseName)
-        .build();
-
-    return databaseInstance;
-  }
-}
