--- lib/features/movies/presentation/controller/i_controller.dart
@@ -1,5 +1,0 @@
-abstract class IController {
-  void initialize();
-
-  void dispose();
-}
