+++ lib/features/movies/presentation/states/data_state.dart
@@ -44,10 +44,3 @@ class DataFailure<T> extends DataState<T> {
           error: error,
         );
 }
-
-class DataEmpty<T> extends DataState<T> {
-  const DataEmpty()
-      : super(
-          state: States.empty,
-        );
-}
