+++ lib/utils/enums/states.dart
@@ -2,5 +2,4 @@ enum States {
   loading,
   success,
   failure,
-  empty,
 }
