+++ lib/utils/providers.dart
@@ -1,38 +1,17 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:http/http.dart';
-import '../features/movies/application/services/implementation/genres_service.dart';
-import '../features/movies/application/services/implementation/movies_service.dart';
-import '../features/movies/application/services/interface/i_service.dart';
-import '../features/movies/data/datasources/local/database_initializer.dart';
-import '../features/movies/data/datasources/local/movies_database.dart';
 import '../features/movies/data/datasources/remote/api_data_source.dart';
 import '../features/movies/data/dto/movie_preview_dto.dart';
-import '../features/movies/data/repository/implementation/database_repository_impl.dart';
-import '../features/movies/data/repository/implementation/genres_repository_impl.dart';
-import '../features/movies/data/repository/implementation/movies_repository_impl.dart';
+import '../features/movies/data/repository/genres_repository_impl.dart';
+import '../features/movies/data/repository/movies_repository_impl.dart';
 import '../features/movies/presentation/controller/movies_controller.dart';
 import 'enums/endpoints.dart';
-
-final moviesDatabaseProvider = Provider<Future<MoviesDatabase>>(
-  (ref) async {
-    DatabaseInitializer dbInitializer = DatabaseInitializer();
-    return dbInitializer.initializeDB();
-  },
-);
-
-final databaseRepository = Provider<IDatabaseRepository>(
-  (ref) => DatabaseRepositoryImpl(
-    ref.watch(
-      moviesDatabaseProvider,
-    ),
-  ),
-);
 
 final clientProvider = Provider<Client>((ref) => Client());
 
 final apiDataSource = Provider<IApiDataSource>(
   (ref) => ApiDataSourceImpl(
-    client: ref.watch(
+    client: ref.read(
       clientProvider,
     ),
   ),
@@ -40,7 +19,7 @@ final apiDataSource = Provider<IApiDataSource>(
 
 final moviesRepository = Provider<IMoviesRepository>(
   (ref) => MoviesRepositoryImpl(
-    apiDataSource: ref.watch(
+    apiDataSource: ref.read(
       apiDataSource,
     ),
   ),
@@ -48,32 +27,16 @@ final moviesRepository = Provider<IMoviesRepository>(
 
 final genresRepository = Provider<IGenresRepository>(
   (ref) => GenresRepositoryImpl(
-    apiDataSource: ref.watch(
+    apiDataSource: ref.read(
       apiDataSource,
     ),
   ),
 );
 
-final moviesService = Provider<IService>(
-  (ref) => MoviesServiceImpl(
-    moviesRepository: ref.watch(moviesRepository),
-    databaseRepository: ref.watch(
-      databaseRepository,
-    ),
-  ),
-);
-
-final genresService = Provider<IService>(
-  (ref) => GenresServiceImpl(
-    genresRepository: ref.watch(genresRepository),
-    databaseRepository: ref.watch(databaseRepository),
-  ),
-);
-
 final moviesControllerProvider = Provider<MoviesController>(
   (ref) => MoviesController(
-    genresService: ref.watch(genresService),
-    moviesService: ref.watch(moviesService),
+    genresRepository: ref.read(genresRepository),
+    moviesRepository: ref.read(moviesRepository),
   ),
 );
